Consolidate related tests into on files
